% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyfilter_function.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply a pixel-level filter function to a bitmap image}
\usage{
apply_filter(bitmap, filter_fn)
}
\arguments{
\item{bitmap}{A 3D raw array of dimensions 3 x width x height.}

\item{filter_fn}{A function that takes a raw RGB pixel (length 3)
and returns a modified raw pixel.}
}
\value{
A modified bitmap in the same format.
}
\description{
This function takes a raw bitmap (as produced by magick)
and applies a filter function to every pixel.
}
\examples{
\dontrun{
img <- magick::image_read("test.jpg")
bmp <- img[[1]]
grey_bmp <- apply_filter(bmp, rgb_to_greyscale)
}

}
