% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{rgb_to_blue_cutoff}
\alias{rgb_to_blue_cutoff}
\title{Apply a blue cutoff filter}
\usage{
rgb_to_blue_cutoff(rgb_vec, cutoff = 127)
}
\arguments{
\item{rgb_vec}{A raw RGB pixel.}

\item{cutoff}{Integer cutoff.}
}
\value{
A raw RGB pixel.
}
\description{
Same logic as other cutoff filters but returns blue.
}
