% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{rgb_to_red_cutoff}
\alias{rgb_to_red_cutoff}
\title{Apply a red cutoff filter to an RGB pixel}
\usage{
rgb_to_red_cutoff(rgb_vec, cutoff = 127)
}
\arguments{
\item{rgb_vec}{A raw RGB pixel.}

\item{cutoff}{Integer between 0 and 255. Default 127.}
}
\value{
A raw vector of length 3 representing the cutoff result.
}
\description{
This function converts a pixel to grayscale, compares its intensity
to a cutoff value, and returns either a red pixel or black.
}
